fix(loader): cast st.Dev to uint64 for cross-platform compatibility
syscall.Stat_t.Dev is int32 on macOS (arm64) but uint64 on Linux, so the
direct assignment to exCacheKey.dev (uint64) fails golangci-lint on the Mac
host. An explicit cast keeps the code correct on both platforms.

+++ internal/loader/load_uprobe.go
@@ -37,7 +37,7 @@ func openExecutable(path string) (*link.Executable, error) {
 		// produce the appropriate error (file not found, etc.).
 		return link.OpenExecutable(path)
 	}
-	key := exCacheKey{dev: st.Dev, ino: st.Ino}
+	key := exCacheKey{dev: uint64(st.Dev), ino: st.Ino}
 	if v, ok := exCache.Load(key); ok {
 		return v.(*link.Executable), nil
 	}
